move ffs rpc to powergate proto, update client
