+++ proto/admin/v1/powergate_admin.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 package proto.admin.v1;
 
-import "ffs/rpc/rpc.proto";
+import "proto/powergate/v1/powergate.proto";
 
 option go_package = "github.com/textileio/powergate/proto/admin/v1";
 
@@ -59,7 +59,7 @@ message QueuedStorageJobsRequest {
 }
 
 message QueuedStorageJobsResponse {
-  repeated ffs.rpc.Job storage_jobs = 1;
+  repeated proto.powergate.v1.Job storage_jobs = 1;
 }
 
 message ExecutingStorageJobsRequest {
@@ -68,7 +68,7 @@ message ExecutingStorageJobsRequest {
 }
 
 message ExecutingStorageJobsResponse {
-  repeated ffs.rpc.Job storage_jobs = 1;
+  repeated proto.powergate.v1.Job storage_jobs = 1;
 }
 
 message LatestFinalStorageJobsRequest {
@@ -77,7 +77,7 @@ message LatestFinalStorageJobsRequest {
 }
 
 message LatestFinalStorageJobsResponse {
-  repeated ffs.rpc.Job storage_jobs = 1;
+  repeated proto.powergate.v1.Job storage_jobs = 1;
 }
 
 message LatestSuccessfulStorageJobsRequest {
@@ -86,7 +86,7 @@ message LatestSuccessfulStorageJobsRequest {
 }
 
 message LatestSuccessfulStorageJobsResponse {
-  repeated ffs.rpc.Job storage_jobs = 1;
+  repeated proto.powergate.v1.Job storage_jobs = 1;
 }
 
 message StorageJobsSummaryRequest {
@@ -95,11 +95,11 @@ message StorageJobsSummaryRequest {
 }
 
 message StorageJobsSummaryResponse {
-  ffs.rpc.JobCounts job_counts = 1;
-  repeated ffs.rpc.Job queued_storage_jobs = 2;
-  repeated ffs.rpc.Job executing_storage_jobs = 3;
-  repeated ffs.rpc.Job latest_final_storage_jobs = 4;
-  repeated ffs.rpc.Job latest_successful_storage_jobs = 5;
+  proto.powergate.v1.JobCounts job_counts = 1;
+  repeated proto.powergate.v1.Job queued_storage_jobs = 2;
+  repeated proto.powergate.v1.Job executing_storage_jobs = 3;
+  repeated proto.powergate.v1.Job latest_final_storage_jobs = 4;
+  repeated proto.powergate.v1.Job latest_successful_storage_jobs = 5;
 }
 
 service PowergateAdminService {
